pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools-git-versioning"]
build-backend = "setuptools.build_meta"

[tool.setuptools-git-versioning]
enabled = true


[project]
name = "django-image-assets"
dynamic = ["version"]
description = "Django application for image assets management"
readme = "README.md"
authors = [
    { name = "Sergey Tikhonov", email = "zimbler@gmail.com" },
]

dependencies = [
    # Django Framework and batteries
    "Django>=2.2,<4.2",
    "django-bitfield~=2.0",
    # Images
    "Pillow>=8.1.2,<10.0.0",
]
license = { text = "MIT" }
keywords = [
    "django",
    "image",
    "assets",
]
classifiers = [
        'Development Status :: 4 - Beta',
        'Environment :: Console',
        'Framework :: Django :: 3.2',
        'Framework :: Django :: 4.0',
        'Framework :: Django :: 4.1',
        'Operating System :: POSIX',
        'Programming Language :: Python :: 3.7',
        'Programming Language :: Python :: 3.8',
        'Programming Language :: Python :: 3.9',
        'Programming Language :: Python :: 3.10',
        'Topic :: Internet :: WWW/HTTP :: Dynamic Content :: Content Management System',
]

[project.urls]
homepage = "https://github.com/just-work/django-image-assets"
documentation = "https://github.com/just-work/django-image-assets/blob/master/README.md"
repository = "https://github.com/just-work/django-image-assets.git"
issues = "https://github.com/just-work/django-image-assets/issues"
